[#138] Refactor: useModalCard 커스텀 훅
